UI Component moves inside src file #22
